Modifications to prevent directory from being over-written by tests
- Not a great way to handle this but only temporary.

+++ tests/globalTeardown.js
@@ -3,7 +3,11 @@ const path = require('path');
 
 module.exports = async() => {
   const dest = path.join(__dirname, '../config/defaults.json');
-  if (fs.existsSync(dest)) {
+  const testStatus = path.join(__dirname, '../config/testStatus.txt');
+  if (fs.existsSync(dest) && fs.existsSync(testStatus)) {
     fs.unlinkSync(dest);
+  }
+  if (fs.existsSync(testStatus)) {
+    fs.unlinkSync(testStatus);
   }
 };
